+++ ev_charge/lib/main.dart
@@ -3,8 +3,8 @@ import 'package:ev_charge/database/database.dart';
 import 'package:ev_charge/database/db_manager.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-//import 'views/calendar_page.dart';
-import 'views/calendar_page2.dart';
+
+import 'views/calendar_page.dart';
 
 /// A global provider for the Drift database
 final dbProvider = Provider<AppDatabase>((ref) {
